backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-generativeai
PyMuPDF
python-docx
reportlab
python-dotenv
pydantic[email]
